requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
celery==5.1.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
kombu==5.1.0
prompt-toolkit==3.0.18
PyJWT==2.1.0
pytz==2021.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
